+++ example/electron3/preload.js
@@ -1,0 +1,5 @@
+const { ipcRenderer } = require("electron");
+
+window.aptabase = {
+  trackEvent: (...args) => ipcRenderer.send("aptabase:trackEvent", ...args),
+};